This fixes a lot of issues talked about in #5
Change config-format to solve some problems.
This commit changes how the config format works.
We take advantage of bash's built-in file sourcing
abilities to allow us to have a config file without
reading a file like before. This solution is more
robust, and it fixes several flaws with the old system
mainly:
  * You can have commands with = in them, something that would break the
  * old system. And you we also no longer break if there is a whitespace
  * in our command :)